Clean EA-MT code
