+++ src/libvgmstream.vcproj
@@ -1623,6 +1623,10 @@
                     >
                 </File>
                 <File
+                    RelativePath=".\coding\ea_mt_decoder_utk.h"
+                    >
+                </File>
+                <File
                     RelativePath=".\coding\ffmpeg_decoder_utils.h"
                     >
                 </File>
